// src/api/stetsom/model/category.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { I18nString } from "./i18nString";
import type { Line } from "./line";

export interface Category {
  id: string;
  name: I18nString;
  slug: I18nString;
  icon_library_id?: string;
  lines: Line[];
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  order: number;
  created_at: string;
  updated_at: string;
}