// src/api/stetsom/model/patchApiCategoriesIdBody.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { PatchApiCategoriesIdBodyName } from "./patchApiCategoriesIdBodyName";
import type { PatchApiCategoriesIdBodySlug } from "./patchApiCategoriesIdBodySlug";

export type PatchApiCategoriesIdBody = {
  name?: PatchApiCategoriesIdBodyName;
  slug?: PatchApiCategoriesIdBodySlug;
  icon_library_id?: string;
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  order?: number;
};